Create mux4to1.v
A multiplexer (MUX) is a combinational circuit that selects one of many input signals and forwards the selected input into a single line. A MUX is often used when a device needs to select between several input sources and transmit a single output to a subsequent stage in a circuit.